Added mutable contact and mutable contact list
Added Mutable Contact within Single Contact, allowing the user to modify the contact shown on the page. This allows them to update the information of the contact without modifying any of the original contact's information prior to clicking the Update button.

Added inputs in Single Contact component to allow for the user to modify the contact information. These updates do not get saved, and the contact information updates back to the original data when selecting the contact again. This data will be used in the future when the Update button gets added.

The HTML for contact information will always show. Inputs and email information is filled with a temporary contact information with the next ID in line. This is set up for the Add Contact button, allowing the user to create a new contact whenever a contact is not selected.

Added nextContactID number passed to the Single Contact component, allowing for future adding of contacts with proper IDs.

Removed React-Router-Dom due to not needing it in the project.

Added "Error" page in the Single Contact component. This should never be hit under any circumstances, but if it does, it informs the user to click on a contact to view the information.

CSS added for Single Contact inputs.