Buttons and Email Adding
Added the ability for the user to add an email. This uses an Input Email button that changes to an Add Email button when clicked. When the user types in an email into the input line and clicks add email, the new email is shown on the page and the Add Email button changes back to the Input Email button.

Added Delete and Save buttons for future use.

CSS for all buttons added.