Adding, Saving, Deleting and Restoring contact information
Added functionality to all the buttons
  - Add button now properly adds a new contact to the database
  - Delete button now properly deletes a contact from the database
  - Save button now properly updates the contact information in the database
  - Cancel button restores the contact's information to its original information. This allows a user to modify the information on the page without actually saving the data, and restoring it if wanted.

Modified how retrieving contacts from the database works. Wrapped the function in the useCallback hook, and added dependancies to useEffect. This lets the application run another retrieval of the contacts from the database whenever the "updateContacts" boolean state is modified. Modification of this state only happens when a user adds, deletes, or modifies contact information.