Validation, Error and Success Messages, New Add Button
Added email validation through regex. Emails with typical formats will be allowed, such as (Test@Test.com, Test.Test@Test.com), and other forms of valid email characters.

Added error messages for when the user inputs a first or last name with no actual text, or an email that is not valid.

Added success messages for Add, Delete, and Save

Added button next to Contacts ( a + sign), to refresh the form to add a new contact.